+++ Sofka.Piguinera.Cotizacion/Sofka.Piguinera.Cotizacion/Models/Strategy/ImplementationDiscountStrategy.cs
@@ -3,12 +3,12 @@
     public class SeniorityOneTwoDiscountStrategy : IDiscountStrategy
     {
         public bool CanApply(int seniority) => seniority >= 1 && seniority <= 2;
-        public decimal Apply() => 0.12m;
+        public float Apply() => 0.12f;
     }
 
     public class SeniorityMoreThanTwoDiscountStrategy : IDiscountStrategy
     {
         public bool CanApply(int seniority) => seniority > 2;
-        public decimal Apply() => 0.17m;
+        public float Apply() => 0.17f;
     }
 }
